refactor(kikan): lift db diagnostics + validate_installation from crates/db [#507]
Second slice of S3.1a (dissolve crates/db → kikan). Moves the
platform-generic database diagnostics — `health_check`,
`diagnose_database` + `DbDiagnostics`, `read_db_runtime_diagnostics` +
`DbRuntimeDiagnostics`, and `validate_installation` — into
`kikan::db::diagnostics`. Callers in `services/api` and the mokumo-db
BDD world now reach for `kikan::db::*` directly; `crates/db` keeps
re-exports under the pre-Stage-3 paths so the remaining crate-shell
tests keep compiling until S3.1b removes the crate.

The unit tests that exercised a migrated database now live in
`mokumo-shop::db::tests`, alongside the other tests that need a fully
initialised profile DB.
